Workaround for setuptools bug
See http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=786977

The dist directory is already cleared prior to the tarball build.
If the version does not match what is expected, force the name
of the tarball in dist/ to be the one we require and warn of doing so.
Also move the unpacked directory to the one required.

+++ share/debian-dev-build.sh
@@ -37,10 +37,18 @@ if [ -d .git ]; then
   LOG=`git log -n1 --pretty=format:"Last change %h by %an, %ar. %s%n" --no-merges`
 fi
 DIR=`mktemp -d`
-mv -v ./dist/${NAME}-${VERSION}.tar.gz ${DIR}/${NAME}_${VERSION}.orig.tar.gz
+if [ -f './dist/${NAME}-${VERSION}.tar.gz' ]; then
+  mv -v ./dist/${NAME}-${VERSION}.tar.gz ${DIR}/${NAME}_${VERSION}.orig.tar.gz
+else
+  echo "WARNING: broken setuptools tarball - Debian bug #786977"
+  mv -v ./dist/${NAME}*.tar.gz ${DIR}/${NAME}_${VERSION}.orig.tar.gz
+fi
 cd ${DIR}
 git clone https://github.com/Linaro/pkg-${NAME}.git
 tar -xzf ${NAME}_${VERSION}.orig.tar.gz
+if [ ! -d ${DIR}/${NAME}-${VERSION} ]; then
+  mv -v ${DIR}/${NAME}-* ${DIR}/${NAME}-${VERSION}
+fi
 cd ${DIR}/pkg-${NAME}/
 dpkg-checkbuilddeps
 git archive master debian | tar -x -C ../${NAME}-${VERSION}
